Expression AST generation now looks correct, some AST/node refactoring
