+++ rust/origen/src/generator/ast.rs
@@ -1,10 +1,8 @@
-use super::processors::ToString;
-use super::stil;
-use crate::generator::processor::*;
+pub use crate::generator::node::Node;
+pub use crate::generator::nodes::Attrs;
 use crate::generator::TestManager;
 use crate::TEST;
 use crate::{Error, Result};
-use num_bigint::BigUint;
 use std::fmt;
 
 #[macro_export]
@@ -21,179 +19,8 @@ macro_rules! node {
     };
 }
 
-#[derive(Clone, Debug, PartialEq, Serialize)]
-pub enum Attrs {
-    // A meta-node type, used to indicate a node who's children should be placed inline at the given location
-    _Inline,
-
-    //////////////////////////////////////////////////////////////////////////////////////////////////////////
-    //// Test (pat gen) nodes
-    //////////////////////////////////////////////////////////////////////////////////////////////////////////
-    Test(String),
-    Comment(u8, String), // level, msg
-    PinWrite(Id, u128),
-    PinVerify(Id, u128),
-    RegWrite(Id, BigUint, Option<BigUint>, Option<String>), // reg_id, data, overlay_enable, overlay_str
-    RegVerify(
-        Id,
-        BigUint,
-        Option<BigUint>,
-        Option<BigUint>,
-        Option<BigUint>,
-        Option<String>,
-    ), // reg_id, data, verify_enable, capture_enable, overlay_enable, overlay_str
-    JTAGWriteIR(u32, BigUint, Option<BigUint>, Option<String>), // size, data, overlay_enable, overlay_str
-    JTAGVerifyIR(
-        u32,
-        BigUint,
-        Option<BigUint>,
-        Option<BigUint>,
-        Option<BigUint>,
-        Option<String>,
-    ), // size, data, verify_enable, capture_enable, overlay_enable, overlay_str
-    JTAGWriteDR(u32, BigUint, Option<BigUint>, Option<String>), // size, data, overlay_enable, overlay_str
-    JTAGVerifyDR(
-        u32,
-        BigUint,
-        Option<BigUint>,
-        Option<BigUint>,
-        Option<BigUint>,
-        Option<String>,
-    ), // size, data, verify_enable, capture_enable, overlay_enable, overlay_str
-    Cycle(u32, bool), // repeat (0 not allowed), compressable
-
-    //// Teradyne custom nodes
-
-    //// Advantest custom nodes
-
-    //////////////////////////////////////////////////////////////////////////////////////////////////////////
-    //// Flow (prog gen) nodes
-    //////////////////////////////////////////////////////////////////////////////////////////////////////////
-    Flow(String),
-
-    //////////////////////////////////////////////////////////////////////////////////////////////////////////
-    //// STIL
-    //////////////////////////////////////////////////////////////////////////////////////////////////////////
-    STIL,
-    STILUnknown,
-    STILVersion(u32, u32), // major, minor
-    STILHeader,
-    STILTitle(String),
-    STILDate(String),
-    STILSource(String),
-    STILHistory,
-    STILAnnotation(String),
-    STILInclude(String, Option<String>),
-    STILSignals,
-    STILSignal(String, stil::SignalType), // name, type
-    STILTermination(stil::Termination),
-    STILDefaultState(stil::State),
-    STILBase(stil::Base, String),
-    STILAlignment(stil::Alignment),
-    STILScanIn(u32),
-    STILScanOut(u32),
-    STILDataBitCount(u32),
-    STILSignalGroups(Option<String>),
-    STILSignalGroup(String),
-    STILSigRefExpr,
-    STILTimeExpr,
-    STILName(String),
-    STILExpr,
-    STILSIUnit(String),
-    STILEngPrefix(String),
-    STILAdd,
-    STILSubtract,
-    STILMultiply,
-    STILDivide,
-    STILParens,
-    STILNumberWithUnit,
-    STILNumber,
-    STILInteger(u64),
-    STILSignedInteger(i64),
-    STILPoint,
-    STILExp,
-    STILMinus,
-    STILPatternExec(Option<String>),
-    STILCategoryRef(String),
-    STILSelectorRef(String),
-    STILTimingRef(String),
-    STILPatternBurstRef(String),
-    STILPatternBurst(String),
-    STILSignalGroupsRef(String),
-    STILMacroDefs(String),
-    STILProcedures(String),
-    STILScanStructuresRef(String),
-    STILStart(String),
-    STILStop(String),
-    STILTerminations,
-    STILTerminationItem,
-    STILPatList,
-    STILPat(String),
-    STILLabel(String),
-    STILTiming(Option<String>),
-    STILWaveformTable(String),
-    STILPeriod,
-    STILInherit(String),
-    STILSubWaveforms,
-    STILSubWaveform,
-    STILWaveforms,
-    STILWaveform,
-    STILWFChar(String),
-    STILEvent,
-    STILEventList(Vec<char>),
-    STILSpec(Option<String>),
-    STILCategory(String),
-    STILSpecItem,
-    STILTypicalVar(String),
-    STILSpecVar(String),
-    STILSpecVarItem(stil::Selector),
-    STILVariable(String),
-    STILSelector(String),
-    STILSelectorItem(String, stil::Selector),
-    STILScanStructures(Option<String>),
-    STILScanChain(String),
-    STILScanLength(u64),
-    STILScanOutLength(u64),
-    STILScanCells,
-    STILScanMasterClock,
-    STILScanSlaveClock,
-    STILScanInversion(u8),
-    STILScanInName(String),
-    STILScanOutName(String),
-    STILNot,
-    STILPattern(String),
-    STILTimeUnit,
-    STILVector,
-    STILCyclizedData,
-    STILNonCyclizedData,
-    STILRepeat(u64),
-    STILWaveformFormat,
-    STILHexFormat(Option<String>),
-    STILDecFormat(Option<String>),
-    STILData(String),
-    STILTimeValue(u64),
-    STILWaveformRef(String),
-    STILCondition,
-    STILCall(String),
-    STILMacro(String),
-    STILLoop(u64),
-    STILMatchLoop(Option<u64>),
-    STILGoto(String),
-    STILBreakPoint,
-    STILIDDQ,
-    STILStopStatement,
-}
-
-impl Node {
-    /// Returns a new node which is the output of the node processed by the given processor.
-    /// Returning None means that the processor has decided that the node should be removed
-    /// from the next stage AST.
-    pub fn process(&self, processor: &mut dyn Processor) -> Result<Option<Node>> {
-        let r = processor.on_node(&self)?;
-        self.process_return_code(r, processor)
-    }
-}
-
+/// An AST provides an API for constructing a node tree, when completed it can be unwrapped
+/// to a node by calling the unwrap() method
 #[derive(Clone)]
 pub struct AST {
     nodes: Vec<Node>,
@@ -201,17 +28,27 @@ pub struct AST {
 
 impl AST {
     /// Create a new AST with the given node as the top-level
-    pub fn new(node: Node) -> AST {
-        AST { nodes: vec![node] }
+    pub fn new() -> AST {
+        AST { nodes: vec![] }
     }
 
+    /// Consumes the AST, converting it to a Node
     pub fn unwrap(&mut self) -> Node {
+        while self.nodes.len() > 1 {
+            let n = self.nodes.pop().unwrap();
+            if let Some(node) = self.nodes.last_mut() {
+                node.add_child(n);
+            }
+        }
         self.nodes.pop().unwrap()
     }
 
     /// Push a new terminal node into the AST
     pub fn push(&mut self, node: Node) {
-        self.nodes.last_mut().unwrap().add_child(node);
+        match self.nodes.last_mut() {
+            Some(n) => n.add_child(node),
+            None => self.nodes.push(node),
+        }
     }
 
     /// Push a new node into the AST and leave it open, meaning that all new nodes
@@ -228,9 +65,10 @@ impl AST {
     /// Close the currently open node
     pub fn close(&mut self, ref_id: usize) -> Result<()> {
         if self.nodes.len() != ref_id {
-            return Err(Error::new(
-                "Attempt to close a parent AST node without first closing all its children, it looks like you have either forgotten to close an open node or given the wrong reference ID"
-            ));
+            return Err(Error::new(&format!(
+                "Attempt to close a parent AST node without first closing all its children (given ID: {}, current length: {}), it looks like you have either forgotten to close an open node or given the wrong reference ID",
+                ref_id, self.nodes.len()
+            )));
         }
         if ref_id == 1 {
             return Err(Error::new("The top-level AST node can never be closed"));
@@ -376,10 +214,7 @@ impl AST {
         }
     }
 
-    // Closes all currently open nodes into a new node but leaving the original state of the AST
-    // unmodified.
-    // This is like a snapshot of the current AST state, mainly useful for printing to the console
-    // for debug.
+    /// Clones the current state of the AST into a Node, leaving the AST unmodified
     pub fn to_node(&self) -> Node {
         let mut node = self.nodes.last().unwrap().clone();
         let num = self.nodes.len();
@@ -421,215 +256,5 @@ impl PartialEq<TEST> for AST {
 impl PartialEq<TestManager> for AST {
     fn eq(&self, test: &TestManager) -> bool {
         self.to_node() == test.to_node()
-    }
-}
-
-type Id = usize;
-
-#[derive(Clone, PartialEq, Serialize)]
-pub struct Node {
-    pub attrs: Attrs,
-    pub meta: Option<Meta>,
-    // This must remain private, potentially we could run across some limitation of this current children
-    // implementation which could force us to change to (for example) an ID based system instead.
-    // Ensuring all interation with this collection is via an API method will allow us to make such a
-    // change under-the-hood without breaking the world.
-    children: Vec<Box<Node>>,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize)]
-pub struct Meta {
-    filename: Option<String>,
-    lineno: Option<usize>,
-}
-
-impl fmt::Display for Node {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.to_string())
-    }
-}
-
-impl fmt::Debug for Node {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.to_string())
-    }
-}
-
-impl PartialEq<AST> for Node {
-    fn eq(&self, ast: &AST) -> bool {
-        *self == ast.to_node()
-    }
-}
-
-impl Node {
-    pub fn new(attrs: Attrs) -> Node {
-        Node {
-            attrs: attrs,
-            children: Vec::new(),
-            meta: None,
-        }
-    }
-
-    pub fn new_with_meta(attrs: Attrs, filename: Option<String>, lineno: Option<usize>) -> Node {
-        Node {
-            attrs: attrs,
-            children: Vec::new(),
-            meta: Some(Meta {
-                filename: filename,
-                lineno: lineno,
-            }),
-        }
-    }
-
-    fn inline(nodes: Vec<Box<Node>>) -> Node {
-        Node {
-            attrs: Attrs::_Inline,
-            meta: None,
-            children: nodes,
-        }
-    }
-
-    pub fn to_string(&self) -> String {
-        ToString::run(self)
-    }
-
-    /// Serializes the AST for import into Python
-    pub fn to_pickle(&self) -> Vec<u8> {
-        serde_pickle::to_vec(self, true).unwrap()
-    }
-
-    pub fn add_child(&mut self, node: Node) {
-        self.children.push(Box::new(node));
-    }
-
-    pub fn insert_child(&mut self, node: Node, offset: usize) -> Result<()> {
-        let len = self.children.len();
-        if offset > len {
-            return Err(Error::new(&format!(
-                "An offset of {} was given to insert a child into a node with only {} children",
-                offset, len
-            )));
-        }
-        let index = self.children.len() - offset;
-        self.children.insert(index, Box::new(node));
-        Ok(())
-    }
-
-    /// Replace the child n - offset with the given node, use offset = 0 to
-    /// replace the last child that was pushed.
-    /// Fails if the node has no children or if the given offset is
-    /// otherwise out of range.
-    pub fn replace_child(&mut self, node: Node, offset: usize) -> Result<()> {
-        let len = self.children.len();
-        if len == 0 {
-            return Err(Error::new(
-                "Attempted to replace a child in a node with no children",
-            ));
-        } else if offset > len - 1 {
-            return Err(Error::new(&format!(
-                "An offset of {} was given to replace a child in a node with only {} children",
-                offset, len
-            )));
-        }
-        let index = self.children.len() - 1 - offset;
-        self.children.remove(index);
-        self.children.insert(index, Box::new(node));
-        Ok(())
-    }
-
-    /// Returns a copy of child n - offset, an offset of 0 means
-    /// the last child that was pushed.
-    /// Fails if the node has no children or if the given offset is
-    /// otherwise out of range.
-    pub fn get_child(&self, offset: usize) -> Result<Node> {
-        let len = self.children.len();
-        if len == 0 {
-            return Err(Error::new(
-                "Attempted to get a child in a node with no children",
-            ));
-        } else if offset > len - 1 {
-            return Err(Error::new(&format!(
-                "An offset of {} was given to get a child in a node with only {} children",
-                offset, len
-            )));
-        }
-        let index = self.children.len() - 1 - offset;
-        Ok(*self.children[index].clone())
-    }
-
-    fn process_return_code(
-        &self,
-        code: Return,
-        processor: &mut dyn Processor,
-    ) -> Result<Option<Node>> {
-        match code {
-            Return::None => Ok(None),
-            Return::ProcessChildren => Ok(Some(self.process_children(processor)?)),
-            Return::Unmodified => Ok(Some(self.clone())),
-            Return::Replace(node) => Ok(Some(node)),
-            // We can't return multiple nodes from this function, so we return them
-            // wrapped in a meta-node and the process_children method will identify
-            // this and remove the wrapper to inline the contained nodes.
-            Return::Unwrap => Ok(Some(Node::inline(self.children.clone()))),
-            Return::Inline(nodes) => Ok(Some(Node::inline(
-                nodes.into_iter().map(|n| Box::new(n)).collect(),
-            ))),
-            Return::InlineBoxed(nodes) => Ok(Some(Node::inline(nodes))),
-        }
-    }
-
-    /// Returns a new node which is a copy of self with its children replaced
-    /// by their processed counterparts.
-    pub fn process_children(&self, processor: &mut dyn Processor) -> Result<Node> {
-        if self.children.len() == 0 {
-            return Ok(self.clone());
-        }
-        let mut nodes: Vec<Box<Node>> = Vec::new();
-        for child in &self.children {
-            if let Some(node) = child.process(processor)? {
-                if let Attrs::_Inline = node.attrs {
-                    for c in node.children {
-                        nodes.push(c);
-                    }
-                } else {
-                    nodes.push(Box::new(node));
-                }
-            }
-        }
-        // Call the end of block handler, giving the processor a chance to do any
-        // internal clean up or inject some more nodes at the end
-        let r = processor.on_end_of_block(&self)?;
-        if let Some(node) = self.process_return_code(r, processor)? {
-            if let Attrs::_Inline = node.attrs {
-                for c in node.children {
-                    nodes.push(c);
-                }
-            } else {
-                nodes.push(Box::new(node));
-            }
-        }
-        Ok(self.replace_children(nodes))
-    }
-
-    /// Returns a new node which is a copy of self with its children replaced
-    /// by the given collection of nodes.
-    pub fn replace_children(&self, nodes: Vec<Box<Node>>) -> Node {
-        let new_node = Node {
-            attrs: self.attrs.clone(),
-            meta: self.meta.clone(),
-            children: nodes,
-        };
-        new_node
-    }
-
-    /// Returns a new node which is a copy of self with its attrs replaced
-    /// by the given attrs.
-    pub fn replace_attrs(&self, attrs: Attrs) -> Node {
-        let new_node = Node {
-            attrs: attrs,
-            meta: self.meta.clone(),
-            children: self.children.clone(),
-        };
-        new_node
     }
 }
