+++ rust/origen/src/generator/mod.rs
@@ -1,5 +1,7 @@
 #[macro_use]
 pub mod ast;
+mod node;
+mod nodes;
 pub mod processor;
 mod processors;
 pub mod stil;
@@ -34,7 +36,8 @@ mod tests {
 
         // Verify comparisons work
 
-        let mut ast1 = AST::new(node!(Test, "trim_vbgap".to_string()));
+        let mut ast1 = AST::new();
+        ast1.push_and_open(node!(Test, "trim_vbgap".to_string()));
         ast1.push(node!(Comment, 1, "Hello".to_string()));
         let r = ast1.push_and_open(node!(RegWrite, 10, 0x12345678_u32.into(), None, None));
         ast1.push(node!(
@@ -56,7 +59,8 @@ mod tests {
         let c = node!(Comment, 1, "Should be outside reg transaction".to_string());
         test.push(c);
 
-        let mut ast2 = AST::new(node!(Test, "trim_vbgap".to_string()));
+        let mut ast2 = AST::new();
+        ast2.push_and_open(node!(Test, "trim_vbgap".to_string()));
         ast2.push(node!(Comment, 1, "Hello".to_string()));
         let r = ast2.push_and_open(node!(RegWrite, 10, 0x12345678_u32.into(), None, None));
         ast2.push(node!(
@@ -82,7 +86,8 @@ mod tests {
 
         let new_ast = test.process(&|ast| UpcaseComments::run(ast).expect("comments upcased"));
 
-        let mut ast = AST::new(node!(Test, "trim_vbgap".to_string()));
+        let mut ast = AST::new();
+        ast.push_and_open(node!(Test, "trim_vbgap".to_string()));
         ast.push(node!(Comment, 1, "HELLO".to_string()));
         let r = ast.push_and_open(node!(RegWrite, 10, 0x12345678_u32.into(), None, None));
         ast.push(node!(
@@ -108,7 +113,8 @@ mod tests {
 
         let new_ast = CycleCombiner::run(&new_ast).unwrap();
 
-        let mut ast = AST::new(node!(Test, "trim_vbgap".to_string()));
+        let mut ast = AST::new();
+        ast.push_and_open(node!(Test, "trim_vbgap".to_string()));
         ast.push(node!(Comment, 1, "HELLO".to_string()));
         let r = ast.push_and_open(node!(RegWrite, 10, 0x12345678_u32.into(), None, None));
         ast.push(node!(
@@ -142,7 +148,8 @@ mod tests {
         test.push(node!(Cycle, 1, true));
         test.push(node!(Cycle, 1, true));
 
-        let mut ast = AST::new(node!(Test, "t1".to_string()));
+        let mut ast = AST::new();
+        ast.push_and_open(node!(Test, "t1".to_string()));
         ast.push(node!(Cycle, 1, false));
         let _r = ast.push_and_open(node!(RegWrite, 10, 0x12345678_u32.into(), None, None));
         ast.push(node!(Cycle, 1, false));
@@ -153,7 +160,8 @@ mod tests {
         // Test replacing the last node
         test.replace(node!(Cycle, 5, false), 0).expect("Ok1");
 
-        let mut ast = AST::new(node!(Test, "t1".to_string()));
+        let mut ast = AST::new();
+        ast.push_and_open(node!(Test, "t1".to_string()));
         ast.push(node!(Cycle, 1, false));
         let _r = ast.push_and_open(node!(RegWrite, 10, 0x12345678_u32.into(), None, None));
         ast.push(node!(Cycle, 1, false));
@@ -164,7 +172,8 @@ mod tests {
         // Test replacing with offset
         test.replace(node!(Cycle, 10, false), 2).expect("Ok2");
 
-        let mut ast = AST::new(node!(Test, "t1".to_string()));
+        let mut ast = AST::new();
+        ast.push_and_open(node!(Test, "t1".to_string()));
         ast.push(node!(Cycle, 1, false));
         let _r = ast.push_and_open(node!(RegWrite, 10, 0x12345678_u32.into(), None, None));
         ast.push(node!(Cycle, 10, false));
@@ -176,7 +185,8 @@ mod tests {
         let test2 = test.clone();
         test2.replace(node!(Cycle, 15, true), 3).expect("Ok3");
 
-        let mut ast = AST::new(node!(Test, "t1".to_string()));
+        let mut ast = AST::new();
+        ast.push_and_open(node!(Test, "t1".to_string()));
         ast.push(node!(Cycle, 1, false));
         ast.push(node!(Cycle, 15, true));
         assert_eq!(test2, ast);
@@ -184,7 +194,8 @@ mod tests {
         // Test replacing an upstream node
         test.replace(node!(Cycle, 15, true), 4).expect("Ok4");
 
-        let mut ast = AST::new(node!(Test, "t1".to_string()));
+        let mut ast = AST::new();
+        ast.push_and_open(node!(Test, "t1".to_string()));
         ast.push(node!(Cycle, 15, true));
         let _r = ast.push_and_open(node!(RegWrite, 10, 0x12345678_u32.into(), None, None));
         ast.push(node!(Cycle, 10, false));
@@ -194,7 +205,8 @@ mod tests {
 
         test.replace(node!(Test, "t2".to_string()), 5).expect("Ok5");
 
-        let mut ast = AST::new(node!(Test, "t2".to_string()));
+        let mut ast = AST::new();
+        ast.push_and_open(node!(Test, "t2".to_string()));
         let _r = ast.push_and_open(node!(RegWrite, 10, 0x12345678_u32.into(), None, None));
         ast.push(node!(Cycle, 10, false));
         ast.push(node!(Cycle, 1, true));
@@ -214,7 +226,8 @@ mod tests {
         test.push(node!(Cycle, 1, true));
         test.push(node!(Cycle, 1, true));
 
-        let mut ast = AST::new(node!(Test, "t1".to_string()));
+        let mut ast = AST::new();
+        ast.push_and_open(node!(Test, "t1".to_string()));
         ast.push(node!(Cycle, 1, false));
         let _r = ast.push_and_open(node!(RegWrite, 10, 0x12345678_u32.into(), None, None));
         ast.push(node!(Cycle, 1, false));
@@ -225,7 +238,8 @@ mod tests {
         // Test inserting the last node
         test.insert(node!(Cycle, 6, false), 0).expect("Ok1");
 
-        let mut ast = AST::new(node!(Test, "t1".to_string()));
+        let mut ast = AST::new();
+        ast.push_and_open(node!(Test, "t1".to_string()));
         ast.push(node!(Cycle, 1, false));
         let _r = ast.push_and_open(node!(RegWrite, 10, 0x12345678_u32.into(), None, None));
         ast.push(node!(Cycle, 1, false));
@@ -237,7 +251,8 @@ mod tests {
         // Test inserting within immediate children
         test.insert(node!(Cycle, 7, false), 2).expect("Ok2");
 
-        let mut ast = AST::new(node!(Test, "t1".to_string()));
+        let mut ast = AST::new();
+        ast.push_and_open(node!(Test, "t1".to_string()));
         ast.push(node!(Cycle, 1, false));
         let _r = ast.push_and_open(node!(RegWrite, 10, 0x12345678_u32.into(), None, None));
         ast.push(node!(Cycle, 1, false));
@@ -249,7 +264,8 @@ mod tests {
 
         test.insert(node!(Cycle, 8, false), 5).expect("Ok2");
 
-        let mut ast = AST::new(node!(Test, "t1".to_string()));
+        let mut ast = AST::new();
+        ast.push_and_open(node!(Test, "t1".to_string()));
         ast.push(node!(Cycle, 1, false));
         let _r = ast.push_and_open(node!(RegWrite, 10, 0x12345678_u32.into(), None, None));
         ast.push(node!(Cycle, 8, false));
@@ -264,7 +280,8 @@ mod tests {
 
         test.insert(node!(Cycle, 9, false), 7).expect("Ok2");
 
-        let mut ast = AST::new(node!(Test, "t1".to_string()));
+        let mut ast = AST::new();
+        ast.push_and_open(node!(Test, "t1".to_string()));
         ast.push(node!(Cycle, 1, false));
         ast.push(node!(Cycle, 9, false));
         let _r = ast.push_and_open(node!(RegWrite, 10, 0x12345678_u32.into(), None, None));
@@ -278,7 +295,8 @@ mod tests {
 
         test.insert(node!(Cycle, 10, false), 9).expect("Ok2");
 
-        let mut ast = AST::new(node!(Test, "t1".to_string()));
+        let mut ast = AST::new();
+        ast.push_and_open(node!(Test, "t1".to_string()));
         ast.push(node!(Cycle, 10, false));
         ast.push(node!(Cycle, 1, false));
         ast.push(node!(Cycle, 9, false));
@@ -316,7 +334,8 @@ mod tests {
             }
         }
 
-        let mut ast = AST::new(node!(Test, "t1".to_string()));
+        let mut ast = AST::new();
+        ast.push_and_open(node!(Test, "t1".to_string()));
         ast.push(node!(Cycle, 1, true));
         let _r = ast.push_and_open(node!(RegWrite, 10, 0x12345678_u32.into(), None, None));
         ast.push(node!(Cycle, 2, true));
